Commit final changes to display flash warning
Add flash warning so that users know they have to
enter a valid url. Add require to URL so that users cant not
submit the field as empty (if they did it would trigger
an error due to the URL extension checker), added formatting
to the date & price displays and user prompts.